chore: extract UI test fixtures and shortlist unit tests
Move ContentView launch-argument fixtures into UITestFixtures so adding
a scenario no longer touches the root view. Split model-shortlist unit
tests out of the 5k-line OpenCodeClientTests file.
